--- packages/react-components/theme-designer/src/utils/themes.ts
@@ -1,7 +1,0 @@
-export type CustomAttributes = {
-  keyColor: string;
-  hueTorsion: number;
-  darkCp: number;
-  lightCp: number;
-  isDark: boolean;
-};
